Remove requirements from README, add iOS note
Remove requirements from README as they are always changing and it's
hard to determine exactly what requirements are. Modern IDEs should be
able to build this without issues and if they can't, something probably
needs to be updated.

Also add a note on iOS. I'm not testing or maintaining anything that's
iOS specific so things might break.